server/app/requirements.txt
boto3
dj-rest-auth
django<4.0
django-allauth
django-cors-headers
django-storages
django-sass-processor
djangorestframework
djangorestframework-simplejwt
channels
channels_redis
asgi-redis
libsass
pillow
psycopg2
pytest_django
redis
uwsgi
